Add class to communicate with QGroundControl.
This commit implements a ZMQMessageHandler class which uses ZeroMQ to
communicate with a custom version of QGroundControl using Google
Protocol Buffers sent over ZeroMQ sockets using a publish-subscribe
model.

Currently, the format of the protocol buffers has yet to be defined, so
the code in this commit does not work. Future work will implement the
protobuf definitions in the secgcs_pb2 module.